Move CalendarContract to CalendarContractWrapper
It seems otherwise that the class is unused and/or stripped out,
possibly by proguard. In any case it seems more explicit.
